mako: Do not try to mount /usb-otg
Starting up TWRP or attempting to do practically any action in it
results in the partitions being mounted and unmounted which in turn
produces an unfriendly "E:Unable to mount '/usb-otg'" log message.
USB OTG is not supported by the device by default as the device is
not able to provide 5V output through the little port. This commit
takes that in mind and removes the mention from the fstab file.

Reference to AOSP for insufficient voltage claim:
https://android.googlesource.com/device/lge/mako/+/fe9f2793424c61588c093df951733347d0d24df4

Change-Id: Iaf96824043ef2eaf718d482c12d7f4b34f3f89f2